fastapi-ocr/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0

# Machine Learning & OCR
torch==2.9.1
torchvision==0.24.1
easyocr==1.7.1
numpy==1.26.4
pillow==10.4.0

# PDF & Document Processing
pdf2image==1.17.0
pdfplumber==0.11.2
mammoth==1.4.19
PyMuPDF==1.24.2
python-docx==1.1.0

# FastAPI File Upload Support
python-multipart==0.0.9

# Extra useful libraries
opencv-python==4.9.0.80
scikit-image==0.22.0
python-dotenv==1.0.1
requests==2.32.3

# Missing dependency added
pandas==2.2.2
openpyxl==3.1.5
